Flask_For_OpenCv/requirements.txt
Flask==2.1.1
Flask_Bcrypt==1.0.1
Flask_Cors==4.0.0
Flask_QRcode==3.2.0
numpy==2.3.3
opencv_contrib_python==4.11.0.86
Pillow==11.3.0
PyMySQL==1.1.0
pyzbar==0.1.9
qrcode==8.2
